refactor: relocate CreateDukanViewModelTest and solve conflict
